refactor[#66] : parsingHeadings , transformArticleData utils,  lib 세그먼트로 변경
