+++ blog/src/entities/article/lib/index.ts
@@ -1,1 +1,2 @@
 export * from "./rehypeMarkdown";
+export * from "./parsingHeadings";
